Add categories presenter
